+++ app/services/projects_for_export_service.rb
@@ -25,11 +25,11 @@ class ProjectsForExportService
   end
 
   private def transfer_projects_by_month_and_year(month, year)
-    Transfer::Project.confirmed.filtered_by_significant_date(month, year)
+    Transfer::Project.filtered_by_significant_date(month, year)
   end
 
   private def conversion_projects_by_month_and_year(month, year)
-    Conversion::Project.confirmed.filtered_by_significant_date(month, year)
+    Conversion::Project.filtered_by_significant_date(month, year)
   end
 
   private def conversion_projects_by_advisory_board_date(month, year)